// libs/shared/page/src/lib/page/media-query/device-types.enum.ts
export enum DeviceTypesEnum {
  COMPUTER = 'computer',
  MOBILE = 'mobile',
  TABLET = 'tablet',
}

export enum BreakpointsEnum {
  XS = '(max-width: 20em)',
  SM = '(max-width: 36em)',
  MD = '(max-width: 48.0625em)',
  LG = '(max-width: 74.5em)',
  XL = '(max-width: 90em)',
  XXL = '(max-width: 104.688em)',
}